Add initial crypto-lib Rust library
This was copied from Sawtooth's Rust SDK (sawtooth-core/sdk/rust). It
provides a signing API which is intended to be pluggable. Currently it
provides a secp256k1 wrapper implementation.

The tests (run with cargo test) provide usage examples. Sawtooth
applications (such as the workload generators) also provide good example
code of this API in use.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>